creaed action type string helper
