+++ packages/core/src/Helpers/createActionTypeString.ts
@@ -1,0 +1,30 @@
+export type ActionTypeString<
+  S extends Record<string, unknown>,
+  Z extends string,
+> = Z extends undefined
+  ? {
+      [K in keyof S as `${string & K}`]: `${string & K}`;
+    }
+  : {
+      [K in keyof S as `[${Z}] - ${string & K}`]: `[${Z}] - ${string & K}`;
+    };
+export const createActionTypeString = <
+  S extends Record<string, unknown>,
+  Z extends string = undefined,
+>(
+  types: S,
+  parentKey?: Z,
+) =>
+  Object.keys(types).reduce((acc, childKey: string) => {
+    const newKey = parentKey ? `[${parentKey}] - ${childKey}` : childKey;
+    return {
+      ...acc,
+      [newKey]: newKey,
+    } as ActionTypeString<S, Z>;
+  }, {} as ActionTypeString<S, Z>);
+
+const parentKey = 'parent';
+const types = { 'send messag': 'send messag', failedme: 'failedme' };
+
+const test = createActionTypeString(types, parentKey);
+const test2 = createActionTypeString(types);
